Add React Navigation autotrack, refactor touchable autotrack to own module
POC React Navigation autotrack
Use different logic
CR + cleanup
