+++ examples/TestDriver/android/app/src/main/java/com/testdriver/MainApplication.java
@@ -3,6 +3,8 @@ package com.testdriver;
 import android.app.Application;
 
 import com.facebook.react.ReactApplication;
+import com.swmansion.gesturehandler.react.RNGestureHandlerPackage;
+import com.heapanalytics.reactnative.RNHeapLibraryPackage;
 import com.facebook.react.ReactNativeHost;
 import com.facebook.react.ReactPackage;
 import com.facebook.react.shell.MainReactPackage;
@@ -24,6 +26,8 @@ public class MainApplication extends Application implements ReactApplication {
     protected List<ReactPackage> getPackages() {
       return Arrays.<ReactPackage>asList(
               new MainReactPackage(),
+            new RNGestureHandlerPackage(),
+            new RNHeapLibraryPackage(),
               new RNHeapLibraryPackage()
       );
     }
