+++ examples/TestDriver/android/settings.gradle
@@ -1,4 +1,8 @@
 rootProject.name = 'TestDriver'
+include ':react-native-gesture-handler'
+project(':react-native-gesture-handler').projectDir = new File(rootProject.projectDir, '../node_modules/react-native-gesture-handler/android')
+include ':@heap_react-native-heap'
+project(':@heap_react-native-heap').projectDir = new File(rootProject.projectDir, '../node_modules/@heap/react-native-heap/android')
 
 include ':app'
